+++ src/test/java/cart/application/OrderServiceTest.java
@@ -1,9 +1,9 @@
 package cart.application;
 
+import cart.application.order.OrderService;
 import cart.domain.*;
 import cart.domain.coupon.Coupon;
 import cart.dto.OrderRequest;
-import cart.dto.OrderResponse;
 import cart.exception.AlreadyUsedCouponException;
 import cart.repository.CartItemRepository;
 import cart.repository.OrderRepository;
@@ -19,7 +19,6 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import java.util.List;
 import java.util.NoSuchElementException;
 
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.mockito.BDDMockito.*;
@@ -118,30 +117,5 @@ class OrderServiceTest {
             assertThatThrownBy(() -> orderService.order(request))
                     .isInstanceOf(NoSuchElementException.class);
         }
-    }
-
-    @Test
-    void 상품을_조회한다() {
-        // given
-        final Member member = new Member(1L, "a@a.com", "1234");
-        final Product chicken = new Product(1L, "치킨", 10000, "imgUrl");
-        final Product dessert = new Product(1L, "desert", 5000, "imgUrl");
-        final OrderItem chickenOrderItem = new OrderItem(chicken, 1);
-        final OrderItem desertOrderItem = new OrderItem(dessert, 1);
-        final Coupon coupon = new Coupon(1L, "1000원 할인 쿠폰", "1000원이 할인 됩니다.", 1000, false);
-
-        final Order order = new Order(List.of(chickenOrderItem, desertOrderItem), member, coupon, chicken.getPrice() + dessert.getPrice());
-        given(orderRepository.findOrderByMemberId(anyLong())).willReturn(List.of(order));
-
-        // when
-        final List<OrderResponse> orderResponses = orderService.findOrderByMember(member);
-
-        // then
-        final OrderResponse result = orderResponses.get(0);
-        assertAll(
-                () -> assertThat(result.getCartItems()).hasSize(2),
-                () -> assertThat(result.getDate()).isNotNull(),
-                () -> assertThat(result.getPrice()).isEqualTo(15000)
-        );
     }
 }
